+++ examples/matmul/matmul.h
@@ -6,6 +6,7 @@ extern "C" {
 #endif
 
 #include <stdint.h>
+#include <stdlib.h>
 
 // Computes matrix multiplication of input matrices given by `input_a` and
 // `input_b`. The output is written at `output`.
@@ -23,27 +24,27 @@ struct matmul_i8_params_t {
   const int8_t* input_a;
   const int8_t* input_b;
   int32_t* output;
-  int16_t row_dim;
-  int16_t inner_dim;
-  int16_t col_dim;
+  size_t row_dim;
+  size_t inner_dim;
+  size_t col_dim;
 };
 
 struct matmul_float_params_t {
   const float* input_a;
   const float* input_b;
   float* output;
-  int16_t row_dim;
-  int16_t inner_dim;
-  int16_t col_dim;
+  size_t row_dim;
+  size_t inner_dim;
+  size_t col_dim;
 };
 
 // This version accepts int8 inputs and int32 outputs.
 void matmul_i8(const int8_t* input_a, const int8_t* input_b, int32_t* output,
-               int16_t row_dim, int16_t inner_dim, int16_t col_dim);
+               size_t row_dim, size_t inner_dim, size_t col_dim);
 
 // This version accepts f32 inputs outputs.
 void matmul_f32(const float* input_a, const float* input_b, float* output,
-                int16_t row_dim, int16_t inner_dim, int16_t col_dim);
+                size_t row_dim, size_t inner_dim, size_t col_dim);
 
 //============================================================================//
 // The functions below are defined only for the testing purposes. All other
@@ -52,11 +53,11 @@ void matmul_f32(const float* input_a, const float* input_b, float* output,
 //============================================================================//
 void matmul_i8_impl_with_params(const struct matmul_i8_params_t* params);
 void matmul_i8_impl(const int8_t* input_a, const int8_t* input_b,
-                    int32_t* output, int16_t row_dim, int16_t inner_dim,
-                    int16_t col_dim);
+                    int32_t* output, size_t row_dim, size_t inner_dim,
+                    size_t col_dim);
 void matmul_f32_impl_with_params(const struct matmul_float_params_t* params);
 void matmul_f32_impl(const float* input_a, const float* input_b, float* output,
-                     int16_t row_dim, int16_t inner_dim, int16_t col_dim);
+                     size_t row_dim, size_t inner_dim, size_t col_dim);
 
 #ifdef __cplusplus
 }  // extern "C"
